Allow tests to run with real ledgers.
To activate this mode, set the environment variable
"LEDGER_LIVE_HARDWARE=1"; this will cause the test suite to access a
real ledger instead of starting a Speculos instance. It does not load
the app onto the ledger, which must be completed as a separate step.

Additionally, GEN_TIME_LIMIT can be set to a number of milliseconds
after which to stop sending any given generative test, and GEN_NUM_RUNS
can be used to set an explicit limit. For instance:

`
nix/env.sh s
LEDGER_LIVE_HARDWARE=1 GEN_TIME_LIMIT=300000 GEN_NUM_RUNS=10 make test
`

will run against a live ledger, and finish any given generative test
after either ten examples or five minutes, whichever comes first.
Running out of time on a generative test does _not_ constitute failure
of the test.

Prompts that were being checked will show the expected prompts in the
console before prompting, and in all cases whether the test suite
expects acceptance or rejection on the ledger is also printed.

+++ hooks.js
@@ -1,7 +1,10 @@
 const SpeculosTransport = require('@ledgerhq/hw-transport-node-speculos').default;
+const HidTransport = require('@ledgerhq/hw-transport-node-hid').default;
 const Avalanche = require('hw-app-avalanche').default;
 const spawn = require('child_process').spawn;
 const fc = require('fast-check');
+var chai = require('chai');
+var { expect, assert } = chai.use(require('chai-bytes'));
 
 const APDU_PORT = 9999;
 const BUTTON_PORT = 8888;
@@ -10,50 +13,81 @@ const AUTOMATION_PORT = 8899;
 exports.mochaHooks = {
   beforeAll: async function () {
     this.timeout(10000); // We'll let this wait for up to 10 seconds to get a speculos instance.
-    speculosProcessOptions=process.env.SPECULOS_DEBUG?{stdio:"inherit"} : {};
-    this.speculosProcess = spawn('speculos', [
+    if(process.env.LEDGER_LIVE_HARDWARE) {
+      this.speculos=await HidTransport.create();
+      this.speculos.button=console.log;
+      console.log(this.speculos);
+    } else {
+      speculosProcessOptions=process.env.SPECULOS_DEBUG?{stdio:"inherit"} : {};
+      this.speculosProcess = spawn('speculos', [
         process.env.LEDGER_APP,
         '--display', 'headless',
         '--button-port', '' + BUTTON_PORT,
         '--automation-port', '' + AUTOMATION_PORT,
         '--apdu-port', '' + APDU_PORT,
       ], speculosProcessOptions);
-    console.log("Speculos started");
-    while (this.speculos === undefined) { // Let the test timeout handle the bad case
-      try {
-        this.speculos = await SpeculosTransport.open({
-          apduPort: APDU_PORT,
-          buttonPort: BUTTON_PORT,
-          automationPort: AUTOMATION_PORT,
-        });
-        if(process.env.DEBUG_BUTTONS) {
-          subButton=this.speculos.button;
-          this.speculos.button=((btns)=>{console.log("Speculos Buttons: " + btns); return subButton(btns);});
+      console.log("Speculos started");
+      while (this.speculos === undefined) { // Let the test timeout handle the bad case
+        try {
+          this.speculos = await SpeculosTransport.open({
+            apduPort: APDU_PORT,
+            buttonPort: BUTTON_PORT,
+            automationPort: AUTOMATION_PORT,
+          });
+          if(process.env.DEBUG_BUTTONS) {
+            subButton=this.speculos.button;
+            this.speculos.button=((btns)=>{console.log("Speculos Buttons: " + btns); return subButton(btns);});
+          }
+        } catch(e) {
+          await new Promise(r => setTimeout(r, 500));
         }
-      } catch(e) {
-        await new Promise(r => setTimeout(r, 500));
       }
     }
     this.ava = new Avalanche(this.speculos, "Avalanche", (_) => { return; });
+    this.flushStderr = function() {
+      if(this.speculosProcess) this.speculosProcess.stdio[2].read();
+    };
   },
   afterAll: async function () {
-    this.speculosProcess.kill();
+    if(this.specuosProcess) {
+      this.speculosProcess.kill();
+    }
   },
   afterEach: async function () {
+    if(this.speculosProcess) {
     stdoutVal=this.speculosProcess.stdio[1].read();
     stderrVal=this.speculosProcess.stdio[2].read();
     if(this.currentTest.state === 'failed') {
       console.log("SPECULOS STDOUT:\n"+stdoutVal);
       console.log("SPECULOS STDERR:\n"+stderrVal);
     }
+    }
   }
 }
 
-async function flowAccept(speculos, acceptPrompt="Accept") {
+async function flowAccept(speculos, expectedPrompts, acceptPrompt="Accept") {
   let promptsPromiseResolve;
   let promptsPromise=new Promise(r => { promptsPromiseResolve = r; });
+  let promptsMatchPromiseResolve;
+  let promptsMatchPromise=new Promise(r => { promptsMatchPromiseResolve = r; });
   let readyPromiseResolve;
   let readyPromise=new Promise(r => { readyPromiseResolve = r; });
+
+  if(!speculos.automationEvents) {
+    if(expectedPrompts) {
+      console.log("Expected prompts: ");
+      for(p in expectedPrompts) {
+        console.log("Prompt %d", p);
+        console.log(expectedPrompts[p][3]);
+        console.log(expectedPrompts[p][17]);
+      }
+    }
+    console.log("Please %s this prompt", acceptPrompt);
+    promptsPromiseResolve();
+    readyPromiseResolve({ prompts: promptsPromise, promptsMatch: Promise.resolve(true) });
+    return readyPromise;
+  }
+
   let isReady = false;
   let isPrimed = false;
   let prompts = [{}];
@@ -63,7 +97,7 @@ async function flowAccept(speculos, acceptPrompt="Accept") {
   // This is so that you can just "await flowAccept(this.speculos);" in a test
   // without actually waiting for the prompts.  If we don't do this, you can
   // end up with two flowAccept calls active at once, causing issues.
-  
+
   await speculos.promptsEndPromise; // Wait for any previous interaction to end.
   speculos.promptsEndPromise=promptsPromise; // Set ourselves as the interaction.
 
@@ -79,7 +113,7 @@ async function flowAccept(speculos, acceptPrompt="Accept") {
         } else {
           if(evt.y === 17 && evt.text === "0.1.0") {
             isReady=true;
-            readyPromiseResolve({prompts: promptsPromise});
+            readyPromiseResolve({prompts: promptsPromise, promptsMatch: promptsMatchPromise});
             return;
           } else {
             speculos.button("Ll");
@@ -116,7 +150,12 @@ async function flowAccept(speculos, acceptPrompt="Accept") {
         } else {
           speculos.button("RLrl");
           subscript.unsubscribe();
-          promptsPromiseResolve(prompts.filter(a=>(Object.keys(a).length!=0))); //.slice(-(n+3), -3));
+          resultingPrompts = prompts.filter(a=>(Object.keys(a).length!=0));
+          if(expectedPrompts) {
+            expect(resultingPrompts).to.deep.equal(expectedPrompts);
+            promptsMatchPromiseResolve( true );
+          }
+          promptsPromiseResolve(resultingPrompts);
         }
       }
     }
@@ -132,9 +171,6 @@ fcConfig={
 	markInterruptAsFailure: false,
   numRuns: parseInt(process.env.GEN_NUM_RUNS || 100)
 };
-
-
-console.log(fcConfig);
 
 fc.configureGlobal(fcConfig);
